libc: Cleanup remaining files
The existing files had some style problems that this patch fixes.
