+++ lib/libc/memset.c
@@ -4,11 +4,14 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <debug.h>
-#include <stdlib.h>
+#include <stddef.h>
 
-void abort(void)
+void *memset(void *dst, int val, size_t count)
 {
-	ERROR("ABORT\n");
-	panic();
+	char *ptr = dst;
+
+	while (count--)
+		*ptr++ = val;
+
+	return dst;
 }
